🚚 api: [CERTIF] move certificationIssueReportSerializer to src/certif/session-mgmt
