+++ api/src/certification/session-management/infrastructure/serializers/certification-issue-report-serializer.js
