+++ api/tests/certification/session-management/unit/infrastructure/serializers/certification-issue-report-serializer_test.js
@@ -1,9 +1,9 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/certification-issue-report-serializer.js';
+import * as serializer from '../../../../../../src/certification/session-management/infrastructure/serializers/certification-issue-report-serializer.js';
 import {
   CertificationIssueReportCategory,
   CertificationIssueReportSubcategories,
-} from '../../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+} from '../../../../../../src/certification/shared/domain/models/CertificationIssueReportCategory.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | certification-issue-report-serializer', function () {
   describe('#serialize()', function () {
